Tons of little fixes. Log panel now shows colors and warning/error outside of the panel again, and now handles the tab character properly. New Profile now has a default name. Changing archive/staging folders no longer tries to copy files to new location. Home now shows all errors if there are path issues. Fixed a bunch of crashes if you have no mods. Modlist tooltip on profiles now only shows on the dropdown button, but shows instantly. Probably fixed an issue where profiles weren't visually updated when you deleted/created a bunch. Prevented user from deleting the active profile. Added border around secondary buttons. Limit paths to one line.

+++ App/main/kotlin/smol_app/cli/SmolCLI.kt
@@ -81,7 +81,7 @@ class SmolCLI(
     }
 
     inner class ModProfileSet : CliktCommand(name = "modprofile-set", help = "Set active mod profile") {
-        val profileId by option().int().required()
+        val profileId by option().required()
 
         override fun run() {
             GlobalScope.launch {
@@ -93,7 +93,7 @@ class SmolCLI(
     }
 
     inner class ModProfileRemove : CliktCommand(name = "modprofile-remove", help = "Remove a mod profile") {
-        val profileId by option().int().required()
+        val profileId by option().required()
 
         override fun run() {
             userManager.removeModProfile(profileId)
